requirements.txt
# Core ML Dependencies
tensorflow>=2.12.0
numpy>=1.23.0
pillow>=9.5.0
matplotlib>=3.7.0
scikit-learn>=1.2.0

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Existing Project Dependencies
firebase-admin>=6.0.0
python-dotenv>=1.0.0
requests>=2.31.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.0

# Optional but Recommended
jupyter>=1.0.0  # For experimentation
seaborn>=0.12.0  # Better visualization
pandas>=2.0.0  # Data analysis
